Crate odbc_ffi

ODBC types those representation is compatible with the ODBC C API.

This layer has not been created using automatic code generation. It is incomplete, i.e. it does not contain every symbol or constant defined in the ODBC C headers. Symbols which are deprecated since ODBC 3 have been left out intentionally. While some extra type safety has been added by grouping some of C's #define constants into enum-types it mostly offers the same power (all) and safety guarantess(none) as the wrapped C-API. ODBC 4.0 is still under development by Microsoft, so these symbols are deactivated by default in the cargo.toml

Reexports

pub use self::SqlDataType::*;

pub use self::SqlCDataType::*;

pub use self::HandleType::*;

pub use self::SqlDriverConnectOption::*;

Enums

Dbc

Env

EnvironmentAttribute
Governs behaviour of EnvironmentAttribute

FetchOrientation
Codes used for FetchOrientation in SQLFetchScroll, SQLDataSources and in SQLDrivers

HandleType
Represented in C headers as SQLSMALLINT

InfoType
Information requested by SQLGetInfo

Obj

SQLRETURN
Indicates the overall success or failure of the function

SqlCDataType

SqlDataType
SQL Data Types

SqlDriverConnectOption
Options for SQLDriverConnect

Stmt

Constants

SQL_MAX_MESSAGE_LENGTH
Maximum message length

SQL_NO_TOTAL

SQL_NTS

SQL_NTSL

SQL_NULL_DATA

SQL_OV_ODBC2

SQL_OV_ODBC3

SQL_OV_ODBC3_80

SQL_SQLSTATE_SIZE

Functions

SQLAllocHandle
Allocates an environment, connection, statement, or descriptor handle.

SQLConnect
SQLConnect establishes connections to a driver and a data source.

SQLDataSources
Returns information about a data source. This function is implemented only by the Driver Manager.

SQLDisconnect

SQLDriverConnect
An alternative to SQLConnect. It supports data sources that require more connection information than the three arguments in SQLConnect, dialog boxes to prompt the user for all connection information, and data sources that are not defined in the system information

SQLDrivers
Lists driver descriptions and driver attribute keywords. This function is implemented only by the Driver Manager.

SQLExecDirect

SQLFetch

SQLFreeHandle
Frees resources associated with a specific environment, connection, statement, or descriptor handle.

SQLGetData

SQLGetDiagRec

SQLGetInfo
Returns general information about the driver and data source associated with a connection

SQLNumResultCols

SQLSetEnvAttr
Sets attributes that govern aspects of environments

SQLTables
Returns the list of table, catalog, or schema names, and table types, stored in a specific data source. The driver returns the information as a result set

Type Definitions

SQLCHAR

SQLHANDLE

SQLHDBC
The connection handle references storage of all information about the connection to the data source, including status, transaction state, and error information.

SQLHENV

SQLHSTMT

SQLHWND

SQLINTEGER

SQLLEN

SQLPOINTER

SQLSMALLINT

SQLUSMALLINT